requirements.txt
flask
flask-cors
torch
torchvision
opencv-python
Pillow
numpy
matplotlib
scipy
